[ESI][Cosim] Reducing MMIO data widths to 32 bit
Some platforms only support 32 bit MMIO, so in accordance with the least
common denominator rule...

+++ include/circt/Dialect/ESI/cosim/CosimDpi.capnp
@@ -56,7 +56,7 @@ struct UntypedData @0xac6e64291027d47a {
 # cases, hardware errors become exceptions.
 interface EsiLowLevel @0xae716100ef82f6d6 {
   # Write to an MMIO register.
-  writeMMIO @0 (address :UInt64, data :UInt64) -> ();
+  writeMMIO @0 (address :UInt32, data :UInt32) -> ();
   # Read from an MMIO register.
-  readMMIO  @1 (address :UInt64) -> (data :UInt64);
+  readMMIO  @1 (address :UInt32) -> (data :UInt32);
 }
